+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix forge ref validation if ref parser is missing [#445](https://github.com/nbbrd/heylogs/issues/445)
+
 ## [0.14.0] - 2025-09-08
 
 This release focuses on improving and extending rules.